scripts/requirements.txt
boto3==1.9.201
botocore==1.12.201
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
docutils==0.14
future==0.17.1
idna==2.8
ijson==2.4
jmespath==0.9.4
pycparser==2.19
pycryptodomex==3.8.2
PyJWT==1.7.1
pyOpenSSL==19.0.0
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
snowflake-connector-python==2.2.10
urllib3==1.25.3
